nmsp_math working with mat3, 4, quat4 andfix previous bug - fix equal() bug

+++ src/core/math/src/nmsp_math/primitive/nmspTuple2.h
@@ -18,7 +18,7 @@ template<class T>
 struct Tuple2_T
 {
 public:
-	using SizeType = size_t;
+	using SizeType = i32;
 
 public:
 	static constexpr size_t s_kElementCount = 2;
