+++ src/core/math/src/nmsp_math/primitive/nmspTuple3.h
@@ -19,7 +19,7 @@ template<class T>
 struct Tuple3_T
 {
 public:
-	using SizeType = size_t;
+	using SizeType = i32;
 
 public:
 	static constexpr size_t s_kElementCount = 3;
