+++ src/core/math/src/nmsp_math/primitive/nmspTuple4.h
@@ -19,7 +19,7 @@ template<class T>
 struct Tuple4_T
 {
 public:
-	using SizeType = size_t;
+	using SizeType = int;
 
 public:
 	static constexpr size_t s_kElementCount = 4;
